+++ saqc/lib/plotting.py
@@ -6,6 +6,8 @@
 
 # -*- coding: utf-8 -*-
 
+from __future__ import annotations
+
 import itertools
 from typing import Optional, Union
 
@@ -58,11 +60,12 @@ def makeFig(
     field: str,
     flags: Flags,
     level: float,
-    max_gap: Optional[str] = None,
-    history: Union[Optional[Literal["valid", "complete"]], list] = "valid",
-    xscope: Optional[slice] = None,
-    phaseplot: Optional[str] = None,
-    ax_kwargs: Optional[dict] = None,
+    max_gap: str | None = None,
+    history: Literal["valid", "complete"] | None | list[str] = "valid",
+    xscope: slice | None = None,
+    phaseplot: str | None = None,
+    ax: mpl.axes.Axes | None = None,
+    ax_kwargs: dict | None = None,
 ):
     """
     Returns a figure object, containing data graph with flag marks for field.
@@ -152,9 +155,10 @@ def makeFig(
         d = _insertBlockingNaNs(d, max_gap)
 
     # figure composition
-    fig = mpl.pyplot.figure(constrained_layout=True, **FIG_KWARGS)
-    grid = fig.add_gridspec()
-    ax = fig.add_subplot(grid[0])
+    if ax is None:
+        fig = mpl.pyplot.figure(constrained_layout=True, **FIG_KWARGS)
+        grid = fig.add_gridspec()
+        ax = fig.add_subplot(grid[0])
 
     _plotVarWithFlags(
         ax,
@@ -172,7 +176,7 @@ def makeFig(
     )
 
     plt.rcParams["font.size"] = default
-    return fig
+    return ax.figure
 
 
 def _plotVarWithFlags(