+++ CHANGELOG.md
@@ -11,12 +11,14 @@ SPDX-License-Identifier: GPL-3.0-or-later
 ### Added
 - add option to not overwrite existing flags to `concatFlags`
 - add option to pass existing axis object to `plot`
+- python 3.11 support
 ### Changed
 - Remove all flag value restrictions from the default flagging scheme `FloatTranslator` 
 - Renamed `TranslationScheme.forward` to `TranslationScheme.toInternal` 
 - Renamed `TranslationScheme.backward` to `TranslationScheme.toExternal` 
 - Changed Default value of the parameter `limit` for `SaQC.interpolateIndex` and `SaQC.interpolateInvalid` to ``None`
 ### Removed
+- python 3.7 support
 ### Fixed
 
 ## [2.2.1](https://git.ufz.de/rdm-software/saqc/-/tags/v2.2.1) - 2022-10-29