+++ CHANGELOG.md
@@ -7,7 +7,14 @@ SPDX-License-Identifier: GPL-3.0-or-later
 # Changelog
 
 ## Unreleased
-[List of commits](https://git.ufz.de/rdm-software/saqc/-/compare/v2.2.1...develop)
+[List of commits](https://git.ufz.de/rdm-software/saqc/-/compare/v2.3.0...develop)
+### Added
+### Changed
+### Removed
+### Fixed
+
+## [2.3.0](https://git.ufz.de/rdm-software/saqc/-/tags/v2.3.0) - 2023-01-17
+[List of commits](https://git.ufz.de/rdm-software/saqc/-/compare/v2.2.1...v2.3.0)
 ### Added
 - add option to not overwrite existing flags to `concatFlags`
 - add option to pass existing axis object to `plot`
@@ -18,6 +25,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 - Renamed `TranslationScheme.backward` to `TranslationScheme.toExternal` 
 - Changed default value of the parameter `limit` for `SaQC.interpolateIndex` and `SaQC.interpolateInvalid` to ``None``
 - Changed default value of the parameter ``overwrite`` for ``concatFlags`` to ``False``
+- Deprecate ``transferFlags`` in favor of ``concatFlags``
 ### Removed
 - python 3.7 support
 ### Fixed